Add author's notebook chapter 5 version 2
I add the author's notebook because in Version 3, there is a lack of images compares to Version 2. The reason is in Version 3, the author decides to not write the explanation in detail, so he does not need to add images. However, I think the images are necessary for the explanation, so I add the author's notebook version 2.